Перенесла порталы для модального окна в компонент Modal
